add corrections of seek

+++ lib/src/chrome_cast_controller.dart
@@ -43,7 +43,7 @@ class ChromeCastController {
   /// If [relative] is set to false sets the video position to an [interval] from the start.
   ///
   /// If [relative] is set to true sets the video position to an [interval] from the current position.
-  Future<void> seek({bool relative = false, double interval = 10.0}) {
+  Future<void> seek({bool relative = false, int interval = 10}) {
     return _chromeCastPlatform.seek(relative, interval, id: id);
   }
